+++ src/main/java/com/example/demo/repository/SeatRepository.java
@@ -8,11 +8,11 @@ import java.util.Optional;
 public interface SeatRepository {
     List<Seat> getAllSeats();
 
-    Optional<Seat> getSeatById(Long id);
+    Optional<Seat> getSeatById(Integer id);
 
     void saveNewSeat(Seat newSeat);
 
-    Optional<Seat> updateSeat(Seat updatedSeat, Long id);
+    Optional<Seat> updateSeat(Seat updatedSeat, Integer id);
 
-    void deleteSeat(Long id);
+    void deleteSeat(Integer id);
 }
